@@ -62,16 +62,16 @@
     "nebular"
   ],
   "dependencies": {
-    "@angular/animations": "^15.2.4",
-    "@angular/cdk": "^15.2.4",
-    "@angular/common": "^15.2.4",
-    "@angular/compiler": "^15.2.4",
-    "@angular/core": "^15.2.4",
-    "@angular/fire": "^7.5.0",
-    "@angular/forms": "^15.2.4",
-    "@angular/platform-browser": "^15.2.4",
-    "@angular/platform-browser-dynamic": "^15.2.4",
-    "@angular/router": "^15.2.4",
+    "@angular/animations": "^16.1.3",
+    "@angular/cdk": "^16.1.3",
+    "@angular/common": "^16.1.3",
+    "@angular/compiler": "^16.1.3",
+    "@angular/core": "^16.1.3",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "^16.1.3",
+    "@angular/platform-browser": "^16.1.3",
+    "@angular/platform-browser-dynamic": "^16.1.3",
+    "@angular/router": "^16.1.3",
     "date-fns": "^2.0.0",
     "docsearch.js": "^2.5.2",
     "eva-icons": "^1.1.3",
@@ -83,38 +83,38 @@
     "pace-js": "^1.0.2",
     "rxfire": "^6.0.0",
     "rxjs": "^6.5.3",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "tslib": "^2.6.0",
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.4",
-    "@angular-devkit/core": "^15.2.4",
-    "@angular-devkit/schematics": "^15.2.4",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "^15.2.4",
-    "@angular/compiler-cli": "^15.2.4",
-    "@angular/language-service": "^15.2.4",
-    "@schematics/angular": "^15.2.4",
+    "@angular-devkit/build-angular": "^16.1.3",
+    "@angular-devkit/core": "^16.1.3",
+    "@angular-devkit/schematics": "^16.1.3",
+    "@angular-eslint/builder": "16.0.3",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/schematics": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "^16.1.3",
+    "@angular/compiler-cli": "^16.1.3",
+    "@angular/language-service": "^16.1.3",
+    "@schematics/angular": "^16.1.3",
     "@types/gulp": "4.0.6",
     "@types/highlight.js": "^9.12.4",
     "@types/jasmine": "~3.10.0",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "^16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.43.0",
+    "@types/node": "^16.18.38",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
     "angular-cli-ghpages": "0.5.0",
     "browserstack-local": "^1.3.7",
     "colors.js": "1.2.4",
     "conventional-changelog-cli": "1.3.4",
     "del": "3.0.0",
     "doc-prsr": "2.2.0",
-    "eslint": "^8.28.0",
+    "eslint": "^8.39.0",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-rxjs": "^4.0.2",
+    "eslint-plugin-rxjs": "^5.0.3",
     "fs-extra": "^8.1.0",
     "gulp": "4.0.2",
     "gulp-replace": "1.0.0",
@@ -131,7 +131,7 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "lint-staged": "^11.2.3",
-    "ng-packagr": "^15.2.2",
+    "ng-packagr": "^16.1.0",
     "npm-run-all": "4.1.5",
     "prettier": "^2.4.1",
     "protractor": "~7.0.0",
@@ -140,7 +140,7 @@
     "stylelint-config-prettier": "^9.0.3",
     "ts-node": "^10.4.0",
     "typedoc": "^0.12.0",
-    "typescript": "~4.9.3"
+    "typescript": "^5.1.6"
   },
   "schematics": "./tools/dev-schematics/dist/collection.json"
 }
