@@ -16,48 +16,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~15.2.4",
-    "@angular/cdk": "~15.2.4",
-    "@angular/common": "~15.2.4",
-    "@angular/compiler": "~15.2.4",
-    "@angular/core": "~15.2.4",
-    "@angular/forms": "~15.2.4",
-    "@angular/platform-browser": "~15.2.4",
-    "@angular/platform-browser-dynamic": "~15.2.4",
-    "@angular/platform-server": "~15.2.4",
-    "@angular/router": "~15.2.4",
+    "@angular/animations": "~16.1.3",
+    "@angular/cdk": "~16.1.3",
+    "@angular/common": "~16.1.3",
+    "@angular/compiler": "~16.1.3",
+    "@angular/core": "~16.1.3",
+    "@angular/forms": "~16.1.3",
+    "@angular/platform-browser": "~16.1.3",
+    "@angular/platform-browser-dynamic": "~16.1.3",
+    "@angular/platform-server": "~16.1.3",
+    "@angular/router": "~16.1.3",
     "@nebular/auth": "file:.lib/auth",
     "@nebular/date-fns": "file:.lib/date-fns",
     "@nebular/eva-icons": "file:.lib/eva-icons",
     "@nebular/moment": "file:.lib/moment",
     "@nebular/security": "file:.lib/security",
     "@nebular/theme": "file:.lib/theme",
-    "@nguniversal/express-engine": "~15.2.0",
+    "@nguniversal/express-engine": "~16.1.1",
     "date-fns": "^2.25.0",
     "eva-icons": "^1.1.3",
     "express": "^4.15.2",
     "moment": "^2.22.2",
     "rxjs": "^7.4.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "tslib": "^2.6.0",
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.2.4",
-    "@angular-eslint/builder": "~15.2.1",
-    "@angular-eslint/eslint-plugin": "~15.2.1",
-    "@angular-eslint/eslint-plugin-template": "~15.2.1",
-    "@angular-eslint/schematics": "~15.2.1",
-    "@angular-eslint/template-parser": "~15.2.1",
-    "@angular/cli": "~15.2.4",
-    "@angular/compiler-cli": "~15.2.4",
-    "@nguniversal/builders": "~15.2.0",
-    "@schematics/angular": "~15.2.4",
+    "@angular-devkit/build-angular": "~16.1.3",
+    "@angular-eslint/builder": "~16.0.3",
+    "@angular-eslint/eslint-plugin": "~16.0.3",
+    "@angular-eslint/eslint-plugin-template": "~16.0.3",
+    "@angular-eslint/schematics": "~16.0.3",
+    "@angular-eslint/template-parser": "~16.0.3",
+    "@angular/cli": "~16.1.3",
+    "@angular/compiler-cli": "~16.1.3",
+    "@nguniversal/builders": "~16.1.1",
+    "@schematics/angular": "~16.1.3",
     "@types/express": "^4.17.0",
     "@types/jasmine": "^3.10.0",
     "@types/node": "^16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.43.0",
-    "@typescript-eslint/parser": "^5.43.0",
-    "eslint": "^8.28.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
+    "eslint": "^8.39.0",
     "jasmine-core": "^3.10.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.0",
@@ -67,7 +67,7 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "protractor": "~7.0.0",
     "ts-node": "^10.4.0",
-    "typescript": "~4.9.5",
+    "typescript": "~5.1.6",
     "webpack-sources": "^3.2.1"
   }
 }
