@@ -24,15 +24,15 @@
     "nebular"
   ],
   "peerDependencies": {
-    "@angular/common": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@angular/forms": "^15.0.0",
+    "@angular/common": "^16.1.3",
+    "@angular/core": "^16.1.3",
+    "@angular/router": "^16.1.3",
+    "@angular/forms": "^16.1.3",
     "@nebular/theme": "11.0.0",
     "rxjs": "^6.5.3 || ^7.4.0"
   },
   "dependencies": {
-    "tslib": "^2.3.0"
+    "tslib": "^2.6.0"
   },
   "sideEffects": false,
   "exports": {
