@@ -1,19 +1,15 @@
-import { Injectable } from '@angular/core';
 import { tap } from 'rxjs/operators';
-import { CanActivate, Router } from '@angular/router';
+import { inject } from '@angular/core';
+import { Router, CanActivateFn } from '@angular/router';
 import { NbAuthService } from '@nebular/auth';
 
-@Injectable()
-export class AuthGuard implements CanActivate {
-  constructor(private authService: NbAuthService, private router: Router) {}
+export const AuthGuard: CanActivateFn = () => {
+  const authService = inject(NbAuthService),
+    router = inject(Router);
 
-  canActivate() {
-    return this.authService.isAuthenticated().pipe(
-      tap((authenticated) => {
-        if (!authenticated) {
-          this.router.navigate(['auth/login']);
-        }
-      }),
-    );
-  }
-}
+  return authService.isAuthenticated().pipe(
+    tap((authenticated) => {
+      if (!authenticated) router.navigate(['auth/login']);
+    }),
+  );
+};
