@@ -7,7 +7,6 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  ComponentFactoryResolver,
   EventEmitter,
   Input,
   OnChanges,
@@ -49,15 +48,12 @@ export class NbCalendarPickerRowComponent<D, T> implements OnChanges {
   // TODO static must be false as of Angular 9.0.0, issues/1514
   @ViewChild(TemplateRef, { read: ViewContainerRef, static: true }) containerRef: ViewContainerRef;
 
-  constructor(private cfr: ComponentFactoryResolver) {}
-
   ngOnChanges() {
-    const factory = this.cfr.resolveComponentFactory(this.component);
-
     this.containerRef.clear();
-
     this.row.forEach((date: D) => {
-      const component = this.containerRef.createComponent(factory, this.containerRef.length);
+      const component = this.containerRef.createComponent(this.component, {
+        index: this.containerRef.length,
+      });
       this.patchWithContext(component.instance, date);
       component.changeDetectorRef.detectChanges();
     });
