@@ -5,15 +5,13 @@
  */
 
 import { ModuleWithProviders, NgModule } from '@angular/core';
-import { CommonModule, DatePipe } from '@angular/common';
+import { DatePipe } from '@angular/common';
 
 import { NB_DATE_ADAPTER, NbDatepickerDirective } from './datepicker.directive';
 import { NbOverlayModule } from '../cdk/overlay/overlay.module';
 import { NbCalendarModule } from '../calendar/calendar.module';
-import { NbCalendarComponent } from '../calendar/calendar.component';
 import { NbDatepickerContainerComponent } from './datepicker-container.component';
 import { NbDatepickerComponent, NbRangepickerComponent, NbBasePickerComponent } from './datepicker.component';
-import { NbCalendarRangeComponent } from '../calendar/calendar-range.component';
 import { NbCalendarRangeModule } from '../calendar/calendar-range.module';
 import { NbDateAdapterService, NbDateTimeAdapterService, NbRangeAdapterService } from './datepicker-adapter';
 import { NbCalendarWithTimeComponent } from './calendar-with-time.component';
@@ -48,12 +46,6 @@ import { NbDateTimePickerComponent } from './date-timepicker.component';
     NbDatepickerComponent,
     NbRangepickerComponent,
     NbBasePickerComponent,
-  ],
-  entryComponents: [
-    NbCalendarComponent,
-    NbCalendarRangeComponent,
-    NbDatepickerContainerComponent,
-    NbCalendarWithTimeComponent,
   ],
 })
 export class NbDatepickerModule {
